Test on Distribution CDF and Bernstein Sampling Implmentation of tests on CDF calculation on a Distribution represented as a Spline, an a simple test on the Bernstein Sampling.
